% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildMetadata.R
\name{BuildMetadata}
\alias{BuildMetadata}
\title{Build gene-snp pairs metadata for eQTL-calling
The input snvmatrix from the user requires the encoding format of 0/1/2/3.}
\usage{
BuildMetadata(
  snvMatrix,
  genedata,
  sparcity = FALSE,
  species = "human",
  group = "celltype",
  snv.number.of.cells = 30,
  expression.min = 1,
  expression.number.of.cells = 30,
  gene_ids = NULL,
  cisDist = NULL
)
}
\arguments{
\item{snvMatrix}{A genotype matrix where each row is one variant and each column is one sample, and the scoring method is 0/1/2/3.}

\item{genedata}{A gene expression matrix where each row is one gene and each column is one sample}

\item{species}{The species that the user wants to select, human or mouse.}

\item{scoring}{The user chooses whether to convert the counting method of the snvMatrix to -1/0/1, TRUE indicates conversion, and FALSE indicates no conversion, default is no conversion.}
}
\description{
Build gene-snp pairs metadata for eQTL-calling
The input snvmatrix from the user requires the encoding format of 0/1/2/3.
}
