% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLcalling.R
\name{eQTLsc}
\alias{eQTLsc}
\title{eQTLsc: Uncover single-cell eQTLs exclusively using scRNA-seq data}
\usage{
eQTLsc(
  genedata,
  snvMatrix = NULL,
  metadata,
  sparcity = FALSE,
  p.adjust.method = "bonferroni",
  useModel = "ZINB",
  p.adjust.Threshold = 0.05
)
}
\arguments{
\item{metadata}{gene和snp的pairs}

\item{p.adjust.method}{p值校验}

\item{expressionMatrix}{基因表达矩阵}

\item{dist}{指定使用负二项分布作为数据的分布类型}

\item{EM}{指定使用EM算法进行参数估计}

\item{theta}{表示模型中过度离散的参数，用于描述计数数据的离散程度}

\item{mu}{表示负二项分布的均值，用于描述非零计数的平均水平}

\item{size}{表示负二项分布的大小参数，用于描述数据的离散程度}

\item{prob}{表示零膨胀分布的概率参数，用于描述计数为零的概率}
}
\value{
A dataframe, each row describes eQTL discovering result of a snv-gene pair
返回一个数据框，每一行描述了一个SNP-gene对的eQTL发现结果
}
\description{
A function designed to identify eQTLs from scRNA-seq data,
utilizing the eQTLsingle_build_metadata function to generate SNP-gene pair information.
这个函数用于从单细胞RNA测序数据中识别eQTLs，并利用BuildMetadata函数生成SNP-gene对信息。
}
