% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeNA.R
\name{ImputeNA}
\alias{ImputeNA}
\title{Title}
\usage{
ImputeNA(expressionMatrix, method = "linear")
}
\arguments{
\item{expressionMatrix}{A dataframe, describes gene expressions, the row names represent gene IDs and the column names represent cell IDs.}

\item{method}{Method for imputating NA for gene expression matrix.}
}
\value{
A expressionMatrix with no missing values.
}
\description{
Title
}
\examples{
# Load test data for cisSceQTL
data(testGene)
# impute the missing values for testGene
expressionMatrix <- ImputeNA(testGene, method = "linear")
}
