% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{Normalize}
\alias{Normalize}
\title{Title}
\usage{
Normalize(expressionMatrix, method = "sctransform")
}
\arguments{
\item{expressionMatrix}{A dataframe, describes gene expressions, the row names represent gene IDs and the column names represent cell IDs.}

\item{method}{Method for normalizing for gene expression dataframe.}
}
\value{
A  normalized gene expression matrix.
}
\description{
Title
}
\examples{
# Load test data for cisSceQTL
data(testGene)
# normalize the data
normalized_data <- Normalize(testGene, method = "sctransform")
}
